new cloud.script module
